chore: rename to ThirdPartySignInButtons
