+++ packages/frontend/src/components/UserSettings/SocialLoginsPanel/index.tsx
@@ -8,7 +8,7 @@ import {
     GoogleLoginButton,
     OktaLoginButton,
     OneLoginLoginButton,
-} from '../../common/GoogleLoginButton';
+} from '../../common/ThirdPartySignInButtons';
 import {
     Bold,
     CardContainer,
